Execute the renewal command only on darwin builds
