+++ orbit/pkg/update/notifications.go
@@ -1,8 +1,6 @@
 package update
 
 import (
-	"fmt"
-	"os/exec"
 	"sync"
 	"time"
 
@@ -65,17 +63,4 @@ func (h *RenewEnrollmentProfileConfigFetcher) GetConfig() (*service.OrbitConfig,
 		}
 	}
 	return cfg, err
-}
-
-func runRenewEnrollmentProfile() error {
-	cmd := exec.Command("/usr/bin/profiles", "renew", "--type", "enrollment")
-	out, err := cmd.CombinedOutput()
-	if err != nil && len(out) > 0 {
-		// just as a precaution, limit the length of the output
-		if len(out) > 512 {
-			out = out[:512]
-		}
-		err = fmt.Errorf("%w: %s", err, string(out))
-	}
-	return err
 }
