+++ orbit/pkg/update/notifications_stub.go
@@ -1,0 +1,7 @@
+//go:build !darwin
+
+package update
+
+func runRenewEnrollmentProfile() error {
+	return nil
+}
